feat: completed the implementation of the marketing pages
