+++ src/domains/external/components/earlyaccess.tsx
@@ -31,7 +31,7 @@ const EarlyAccess = () => {
   };
 
   return (
-    <div className="w-full bg-[#520052]">
+    <div className="w-full bg-[#520052] bg-cover bg-center" style={{ backgroundImage: "url('/Map.png')" }}>
       <div className="grid grid-cols-1 md:grid-cols-2">
         {/* Left Column: Image */}
         <div className="flex items-center justify-center">
